# Stop duplicating metrics attrSet