@carsonip 2023-07-26
The alloc before this PR was due to metric.attrOpt escaping:
```
aggregators/writer.go:133:40: ... argument escapes to heap
aggregators/writer.go:133:72: metric.attrOpt{...} escapes to heap
```

A nil sortable doesn't work. And if we pass in a sortable to NewSetWithSortable, we'll have to allocate it ourselves anyway. So we might as well use the otel pooled sortable.